Removing API V3 support and Flake8 cleanup
