+++ tests/unit_tests/local_plugins/hello.py
@@ -22,9 +22,11 @@ def hello_reply_formatting(message):
 def hello_send(message):
     message.send('hello channel!')
 
+
 @listen_to('hello$')
 def hello_send_alternative(message):
     message.send('hello channel!')
+
 
 @listen_to('hello_decorators')
 @respond_to('hello_decorators')
@@ -56,4 +58,3 @@ def hello_comment(message):
 @listen_to('hello_react', re.IGNORECASE)
 def hello_react(message):
     message.react(':+1:')
-
