+++ tests/unit_tests/single_plugin/mock_plugin.py
@@ -2,6 +2,7 @@ from mmpy_bot.utils import allow_only_direct_message
 from mmpy_bot.utils import allowed_users
 from mmpy_bot.bot import respond_to
 
+
 @respond_to('^admin$')
 @allow_only_direct_message()
 @allowed_users('admin', 'root')
